refactor(multiple): update to typescript 3.7.5 move to dist vs lib other minor
