+++ security/symmetric/.npmignore
