+++ security/symmetric/index.d.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { SymmetricKeySecurityClient } from './lib/symmetric_key';
+export { SymmetricKeySecurityClient } from './dist/symmetric_key';
