+++ security/tpm/index.d.ts
@@ -1,5 +1,5 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { TpmSecurityClient } from './lib/tpm';
-export { TpmAuthenticationProvider } from './lib/tpm_authentication_provider';
+export { TpmSecurityClient } from './dist/tpm';
+export { TpmAuthenticationProvider } from './dist/tpm_authentication_provider';
