+++ security/tpm/index.js
@@ -4,6 +4,6 @@
 'use strict';
 
 module.exports = {
-  TpmSecurityClient : require('./lib/tpm').TpmSecurityClient,
-  TpmAuthenticationProvider: require('./lib/tpm_authentication_provider').TpmAuthenticationProvider
+  TpmSecurityClient : require('./dist/tpm').TpmSecurityClient,
+  TpmAuthenticationProvider: require('./dist/tpm_authentication_provider').TpmAuthenticationProvider
 };
