+++ security/tpm/package.json
@@ -24,7 +24,7 @@
     "source-map-support": "^0.5.16",
     "ts-node": "^8.6.2",
     "tslint": "^6.1.0",
-    "typescript": "2.9.2"
+    "typescript": "3.7.5"
   },
   "scripts": {
     "npmlockrefresh": "npm i --package-lock-only",
