+++ security/x509/index.d.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { X509Security } from './lib/x509';
+export { X509Security } from './dist/x509';
