+++ security/x509/tsconfig.json
@@ -1,10 +1,12 @@
 {
     "compilerOptions": {
-        "outDir": "./lib",
+        "outDir": "./dist",
         "target":"es5",
         "sourceMap": true,
         "declaration": true,
-        "noUnusedLocals": true
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "removeComments": false
     },
     "include": [
         "./src/**/*.ts"
