+++ service/iothub.d.ts
@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { Client } from './lib/client';
-export import ConnectionString = require('./lib/connection_string');
-export { Registry } from './lib/registry';
-export import SharedAccessSignature = require('./lib/shared_access_signature');
-export { Amqp } from './lib/amqp';
-export { AmqpWs } from './lib/amqp_ws';
-export { DeviceMethodParams } from './lib/interfaces';
-export { JobClient } from './lib/job_client';
-export * from './lib/configuration';
-export { Device } from './lib/device';
-export { Module } from './lib/module';
-export { Twin } from './lib/twin';
+export { Client } from './dist/client';
+export import ConnectionString = require('./dist/connection_string');
+export { Registry } from './dist/registry';
+export import SharedAccessSignature = require('./dist/shared_access_signature');
+export { Amqp } from './dist/amqp';
+export { AmqpWs } from './dist/amqp_ws';
+export { DeviceMethodParams } from './dist/interfaces';
+export { JobClient } from './dist/job_client';
+export * from './dist/configuration';
+export { Device } from './dist/device';
+export { Module } from './dist/module';
+export { Twin } from './dist/twin';
