+++ service/package.json
@@ -27,7 +27,7 @@
     "source-map-support": "^0.5.16",
     "ts-node": "^8.6.2",
     "tslint": "^6.1.0",
-    "typescript": "2.9.2",
+    "typescript": "3.7.5",
     "uuid": "^3.3.2"
   },
   "scripts": {
