+++ service/test/_client_test.js
@@ -6,8 +6,8 @@
 var assert = require('chai').assert;
 var sinon = require('sinon');
 var EventEmitter = require('events').EventEmitter;
-var Amqp = require('../lib/amqp.js').Amqp;
-var Client = require('../lib/client.js').Client;
+var Amqp = require('../dist/amqp.js').Amqp;
+var Client = require('../dist/client.js').Client;
 var Message = require('azure-iot-common').Message;
 var errors = require('azure-iot-common').errors;
 var SimulatedAmqp = require('./amqp_simulated.js');
