+++ service/test/_device_test.js
@@ -3,7 +3,7 @@
 
 'use strict';
 var assert = require('chai').assert;
-var Device = require('../lib/device.js').Device;
+var Device = require('../dist/device.js').Device;
 
 describe('Device', function() {
   describe('constructor', function() {
