+++ service/tsconfig.json
@@ -1,16 +1,18 @@
 {
     "compilerOptions": {
-        "outDir": "./lib",
+        "outDir": "./dist",
         "target":"es5",
         "sourceMap": true,
         "declaration": true,
-        "noUnusedLocals": true
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "removeComments": false
     },
     "include": [
         "./src/**/*.ts"
     ],
     "exclude": [
-        "./lib",
+        "./dist",
         "./samples"
     ]
 }
