+++ ts-e2e/package.json
@@ -7,7 +7,7 @@
     "npmlockrefresh": "npm i --package-lock-only",
     "lint": "tslint --project . -c ../tslint.json",
     "build": "tsc",
-    "test": "mocha ./lib/*.tests.js"
+    "test": "mocha ./dist/*.tests.js"
   },
   "keywords": [],
   "author": "Microsoft Corporation",
@@ -29,7 +29,7 @@
     "debug": "^4.1.1",
     "mocha": "^7.1.1",
     "tslint": "^6.1.0",
-    "typescript": "2.9.2",
+    "typescript": "3.7.5",
     "uuid": "^3.3.2",
     "uuid-buffer": "^1.0.3"
   }
