+++ ts-e2e/tsconfig.json
@@ -1,10 +1,12 @@
 {
   "compilerOptions": {
-    "outDir": "lib",
+    "outDir": "dist",
     "target": "es2015",
     "sourceRoot": "src",
     "sourceMap": true,
     "moduleResolution": "node",
-    "noUnusedLocals": true
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "removeComments": false
   }
 }
